Implemented transparent unique class identifiers

+++ ue35/Umsetzungskomponente_02/src/main/java/at/ac/tuwien/mnsa/message/MessageFactory.java
@@ -3,5 +3,5 @@ package at.ac.tuwien.mnsa.message;
 import java.io.Serializable;
 
 public interface MessageFactory<T extends Serializable> {
-	public Message<T> create(int type, byte[] payload) throws MessageCreationException;
+	public Message<T> create(byte type, byte[] payload) throws MessageCreationException;
 }
